+++ test-fixtures/packages/template-standard/include/{{node_name_cpp}}.hpp
@@ -1,0 +1,1 @@
+#ifndef {{HEADER_GUARD}}_NODE_HPP\n#define {{HEADER_GUARD}}_NODE_HPP\n#endif